legg til warning hvis skjema ikke er sendt in

+++ src/main/java/no/nav/permitteringsskjemaapi/replay/ReplayService.java
@@ -4,13 +4,21 @@ import no.nav.permitteringsskjemaapi.integrasjon.arbeidsgiver.Permitteringsskjem
 import no.nav.permitteringsskjemaapi.permittering.Permitteringsskjema;
 import no.nav.permitteringsskjemaapi.permittering.PermitteringsskjemaRepository;
 import no.nav.permitteringsskjemaapi.permittering.domenehendelser.PermitteringsskjemaSendtInn;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Profile;
+import org.springframework.data.domain.Pageable;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
+
+import javax.transaction.Transactional;
+import java.lang.invoke.MethodHandles;
 
 @Profile({"dev-gcp", "prod-gcp"})
 @Service
 public class ReplayService {
+
+    private final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
 
     private final ReplayRepository replayRepository;
     private final PermitteringsskjemaRepository permitteringsskjemaRepository;
@@ -26,12 +34,17 @@ public class ReplayService {
         this.permitteringsskjemaProdusent = permitteringsskjemaProdusent;
     }
 
+    @Transactional
     @Scheduled(fixedRate = 5_000)
     public void scheduleFixedRateTask() {
-        replayRepository.fetchReplayQueueItems().forEach(replayQueueItem -> {
+        replayRepository.fetchReplayQueueItems(Pageable.ofSize(10)).forEach(replayQueueItem -> {
             Permitteringsskjema skjema = permitteringsskjemaRepository.findById(replayQueueItem.skjemaId).orElseThrow();
-            permitteringsskjemaProdusent.sendInn(new PermitteringsskjemaSendtInn(skjema, "replayer"));
-            replayRepository.delete(replayQueueItem);
+            if (skjema.getSendtInnTidspunkt() == null) {
+                log.warn("skjema er ikke sendt inn, skipper replay av skjemaId {}", skjema.getId());
+            } else {
+                permitteringsskjemaProdusent.sendInn(new PermitteringsskjemaSendtInn(skjema, "replayer"));
+                replayRepository.delete(replayQueueItem);
+            }
         });
     }
 }
